Initialize filesystem at image-build
Previously, the image was initialized at the startup of the namenode
service.  This caused a race condition at runtime if the datanode tried
to connect to the namenode before the initialization was complete.

Instead, perform the initialization while building the image.

This works because of Docker's behavior when mounting a volume for the
first time.  When Docker mounts a given volume into a container for the
first time, and the directory in the container has files, those files
are copied into the volume.