Make initialization script idempotent
Previously the initialization script would fail trying to ask the TTY
whether it *really* wanted to reformat the HDFS.  The presence of the
nonInteractive flag indicates that if the HDFS is already formatted then
no action should be taken.

This change is necessary for the initialization to take place in a
Kubernetes InitContainer.  Unlike Docker Swarm, Kubernetes does not copy
the container contents into a new volume when mounting it for the first
time, so we cannot initialize the HDFS in the image.